+++ src/qtile_lxa/widget/theme/bar/__init__.py
@@ -1,12 +1,3 @@
-from .decorator.bar_decorator import DecoratedBar
-from .decorator.decoration_switcher import DecorationChanger
-from .splitter.split_mode_switcher import BarSplitModeChanger
-from .transparency.transparency_mode_swicher import BarTransparencyModeChanger
+from .bar_decorator import DecoratedBar
 
-
-__all__ = [
-    "DecoratedBar",
-    "DecorationChanger",
-    "BarSplitModeChanger",
-    "BarTransparencyModeChanger",
-]
+__all__ = ["DecoratedBar"]
