+++ src/qtile_lxa/widget/theme/bar/bar_decorator.py
