+++ src/qtile_lxa/widget/theme/controllers/bar_decoration.py
