+++ src/qtile_lxa/widget/theme/controllers/bar_spliter.py
