+++ src/qtile_lxa/widget/theme/controllers/color_rainbow.py
