--- src/qtile_lxa/widget/theme2/__init__.py
@@ -1,11 +1,0 @@
-from . import config, controllers, utils, bar
-from . import manager
-
-
-__all__ = [
-    "config",
-    "controllers",
-    "utils",
-    "bar",
-    "manager",
-]
