--- src/qtile_lxa/widget/theme2/bar/__init__.py
@@ -1,3 +1,0 @@
-from .bar_decorator import DecoratedBar
-
-__all__ = ["DecoratedBar"]
