--- src/qtile_lxa/widget/theme2/utils/__init__.py
@@ -1,3 +1,0 @@
-from . import colors
-
-__all__ = ["colors"]
