+++ src/qtile_lxa/widget/theme_old/__init__.py
@@ -1,0 +1,13 @@
+from .theme_manager import theme_manager
+from . import bar, color, config, pywall, utils, vidwall
+
+
+__all__ = [
+    "bar",
+    "color",
+    "config",
+    "pywall",
+    "utils",
+    "vidwall",
+    "theme_manager",
+]
