+++ src/qtile_lxa/widget/theme_old/bar/__init__.py
@@ -1,0 +1,12 @@
+from .decorator.bar_decorator import DecoratedBar
+from .decorator.decoration_switcher import DecorationChanger
+from .splitter.split_mode_switcher import BarSplitModeChanger
+from .transparency.transparency_mode_swicher import BarTransparencyModeChanger
+
+
+__all__ = [
+    "DecoratedBar",
+    "DecorationChanger",
+    "BarSplitModeChanger",
+    "BarTransparencyModeChanger",
+]
