+++ src/qtile_lxa/widget/theme_old/bar/decorator/decoration_switcher.py
