+++ src/qtile_lxa/widget/theme_old/pywall/sources/git.py
@@ -1,7 +1,7 @@
 import subprocess
 from pathlib import Path
 from libqtile.log_utils import logger
-from qtile_lxa.widget.theme.config import ThemeConfig
+from qtile_lxa.widget.theme_old.config import ThemeConfig
 from qtile_lxa.utils.notification import send_notification
 from qtile_lxa.utils.process_lock import ProcessLocker
 from .utils import sync_config_for_source
